pyproject.toml

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "steelcity-impact-bridge"
version = "0.1.0"
description = "Raspberry Pi BLE bridge for steel-plate impact detection (AMG T0 + WitMotion BT50 PoC)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "bleak>=0.22.0",
  "numpy>=1.24",
  "cbor2>=5.6.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0","ruff>=0.5.0","black>=24.0.0"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.setuptools.packages.find]
where = ["src"]
